Tile Bleeder v1.1
Updated tile bleeder to use a more efficient process to produce faster and cleaner results.

Changes include:
- Updated PerformBleed to now find the nearest relative pixel to draw its bleed from, instead of scanning all cardinal directions for each pixel.
- Updated how PerformBleed handles depth. Before, the method would be called N number of time as the user specifies, thus allowing the bleed to iterate on itself. Not only was this slow as it would have to iterate over each pixel in the image N number of times, but it would also produce small imperfections the more iterates you've requested.
Now PerformBleed will find the nearest pixel to draw from the base, and if it is within the specified depth or range, it will apply that pixel, or it will continue. The resulting formula is faster than the previous publication as the image only needs to be scanned once. It also creates a cleaner result due to all bleeds being drawn from the same base image, rather than being drawn from the a series of bled images.
- The unprotected method allows the use of the older method if iterations and is intended for ununiform sprites, but its results are slower to produce, and are prone to mutations. Be warned.